authUpdates

+++ src/controllers/ProjectController.ts
@@ -4,6 +4,8 @@ import Project from '../models/Project'
 export class ProjectController {
     static createProject = async (req: Request, res: Response) => {
         const project = new Project(req.body)
+        project.manager = req.user.id
+        
         try {
             await project.save()
             res.send('Proyecto Creado Correctamente')
@@ -14,7 +16,11 @@ export class ProjectController {
 
     static getAllProjects = async (req: Request, res: Response) => {
         try {
-            const projects = await Project.find({})
+            const projects = await Project.find({
+                $or: [
+                    {manager: {$in: req.user.id}}
+                ]
+            })
             res.json(projects)
         } catch (error) {
             console.log(error);
@@ -28,6 +34,12 @@ export class ProjectController {
             const project = await Project.findById(id).populate('tasks')
             if(!project) {
                 const error = new Error('Proyecto no encontrado')
+                res.status(404).json({error: error.message})
+                return
+            }
+
+            if(project.manager.toString() !== req.user.id.toString()) {
+                const error = new Error('Acción no válida')
                 res.status(404).json({error: error.message})
                 return
             }
@@ -47,6 +59,13 @@ export class ProjectController {
                 res.status(404).json({error: error.message})
                 return
             }
+
+            if(project.manager.toString() !== req.user.id.toString()) {
+                const error = new Error('Solo el Manager puede actualizar el proyecto')
+                res.status(404).json({error: error.message})
+                return
+            }
+
             project.projectName = req.body.projectName
             project.clientName = req.body.clientName
             project.description = req.body.description
@@ -67,6 +86,13 @@ export class ProjectController {
                 res.status(404).json({error: error.message})
                 return
             }
+
+            if(project.manager.toString() !== req.user.id.toString()) {
+                const error = new Error('Solo el Manager puede eliminar el proyecto')
+                res.status(404).json({error: error.message})
+                return
+            }
+
             await project.deleteOne()
             res.send('Proyecto Eliminado')
